+++ e2e/module-name-mapper-mock/src/storage/track/Track.js
@@ -1,7 +1,4 @@
 module.exports.Track = class Track {
-  constructor() {
-    /**/
-  }
   someRandomFunction() {
     return 42;
   }