+++ packages/jest-runtime/src/__tests__/runtime_gen_mock_from_module.test.js
@@ -46,7 +46,7 @@ describe('Runtime', () => {
         const mockModule = root.jest.genMockFromModule(
           'module/name/genMockFromModule',
         );
-        console.log(mockModule);
+
         expect(mockModule.test.mock).toBeTruthy();
       }));
   });